@@ -1,17 +1,16 @@
 using ProductService.Model;
 
-namespace ProductService.DTOs
+namespace ProductService.DTOs;
+
+public class ProductInfo
 {
-    public class ProductInfo
-    {
-        public int Id { get; set; }
-        public string Name { get; set; } = string.Empty;
-        public int CategoryId { get; set; }
-        public string CategoryName { get; set; } = string.Empty;
-        public decimal UnitPrice { get; set; }
-        public int StockLevel { get; set; }
-        public List<UserComment> Comments { get; set; } = new List<UserComment>();
-        public List<Photo> Photos { get; set; } = new List<Photo>();
-        public bool InStock { get; set; }
-    }
+    public int Id { get; set; }
+    public string Name { get; set; } = string.Empty;
+    public int CategoryId { get; set; }
+    public string CategoryName { get; set; } = string.Empty;
+    public decimal UnitPrice { get; set; }
+    public int StockLevel { get; set; }
+    public List<UserComment> Comments { get; set; } = new List<UserComment>();
+    public List<Photo> Photos { get; set; } = new List<Photo>();
+    public bool InStock { get; set; }
 }
